Move oak::wasm module to separate file
Also include the SPACE_BYTE_PER_HANDLE constant, as it forms part of
the ABI.  Make the module public too.
